requirements.txt
# Core LLM and RAG
langchain
langgraph
langchain-google-genai
chromadb

# PDF/Text handling
PyMuPDF
python-dotenv

# Backend
fastapi
uvicorn

# Frontend (optional now, needed next step)
streamlit

# Typing + API
pydantic

# CORS
httpx

# NGROk
ngrok